+++ packages/user/FractalCore/ui/src/pages/fractal/governance/judicial.tsx
@@ -48,7 +48,7 @@ export const Judicial = () => {
     }
 
     return (
-        <div className="mx-auto w-full max-w-5xl space-y-8 p-4 px-6">
+        <div className="mx-auto w-full max-w-5xl space-y-6 p-4 px-6">
             <ExileFractalMemberModal
                 openChange={setShowModal}
                 show={showModal}
