+++ packages/user/FractalCore/ui/src/pages/guilds/leadership/components/register-candidacy-modal.tsx
