+++ packages/user/FractalCore/ui/src/pages/guilds/membership/my-guild-membership.tsx
@@ -37,9 +37,6 @@ export const MyGuildMembership = () => {
 
     return (
         <div className="mx-auto w-full max-w-5xl p-4 px-6">
-            <div className="flex h-9 items-center">
-                <h1 className="text-lg font-semibold">My Guild Membership</h1>
-            </div>
             <div className="mt-3 space-y-6">
                 {isLoading ? (
                     <>
